scraper/requirements.txt
crawl4ai
pandas>=1.3.0
pyyaml>=5.4.1
numpy>=1.21.0
python-json-logger>=2.0.2
